Fixed the build using labels with test
form.lable not only creates a <label> tag for the associated <input>
and/or <textarea> that is to be filled in, but also has the attribute
for="<model>_<filed>" which matches the id of the releveant input tag.
This leads to form behaviour like clicking on the lable field, puts the
cursor into the appropriate input filed etc.

to find out what the interface for <% form.label :filed_name ... is
searching "rails form.label" led me to
https://guides.rubyonrails.org/form_helpers.html which very quickly
showed me an example that a first argument of string will become the
lable text, ie
  before:
    <% form.label, :field_name, class: ...
    Filed name
  after
    <% form.label, :field_name, "cuStom naMe", class: ...
    cuStom naMe